refactor: NewWorksFromGalleriesYouFollow section
